feat: split vector_style into fill_style and stroke_style
